+++ Module.php
@@ -1,25 +1,3 @@
 <?php
-namespace TwbBundle;
-class Module implements
-	\Zend\ModuleManager\Feature\ConfigProviderInterface,
-	\Zend\ModuleManager\Feature\AutoloaderProviderInterface{
 
-	/**
-	 * @see \Zend\ModuleManager\Feature\AutoloaderProviderInterface::getAutoloaderConfig()
-	 * @return array
-	 */
-	public function getAutoloaderConfig(){
-        return array(
-            'Zend\Loader\ClassMapAutoloader' => array(
-                __DIR__.DIRECTORY_SEPARATOR.'autoload_classmap.php'
-            )
-        );
-    }
-
-    /**
-     * @return array
-     */
-    public function getConfig(){
-        return include __DIR__.DIRECTORY_SEPARATOR.'config/module.config.php';
-    }
-}
+require_once __DIR__ . '/src/TwbBundle/Module.php';
